+++ example/demo/analysis-new/index.less
@@ -1,7 +1,7 @@
-.scene-4-new {
+.analysis-new {
   width: 100%;
   height: 100%;
-  .scene-4-new-canvas {
+  .analysis-new-canvas {
     width: 100%;
     height: 100%;
   }
