+++ example/demo/analysis-new/node.less
@@ -1,4 +1,4 @@
-.scene-4-base-node {
+.analysis-base-node {
   position: absolute;
   height: 36px;
   width: 116;
@@ -38,7 +38,7 @@
     background: #9254DE;
   }
 }
-.scene-4-circle-base-node {
+.analysis-circle-base-node {
   position: absolute;
   background: #FFFFFF;
   border: 1px solid #D9D9D9;
