+++ example/demo/decision-new/data.js
@@ -7,345 +7,315 @@ module.exports = {
   nodes: [
     {
       id: '1',
-      label: '开始',
-      className: 'green',
-      iconType: 'icon-bofang',
-      top: 50,
-      left: 63,
-      Class: Node
+      label: 'Number',
+      className: 'purple',
+      iconType: 'icon-wo',
+      top: 25,
+      left: 380,
+      Class: Node,
+      endpoints: [{
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '2',
-      label: '默认通过',
-      className: 'simple-blue',
-      iconType: 'icon-rds',
-      top: 150,
-      left: 50,
-      Class: Node
+      label: 'Name',
+      className: 'purple',
+      iconType: 'icon-wo',
+      top: 25,
+      left: 520,
+      Class: Node,
+      endpoints: [{
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '3',
-      label: '政策监管准入',
-      className: 'deep-blue',
-      iconType: 'icon-guize-kai',
-      top: 250,
-      left: 36,
+      label: 'Skills',
+      className: 'purple',
+      iconType: 'icon-wo',
+      top: 25,
+      left: 660,
       Class: Node,
-      endpoints: [
-        {
-          id: 'right',
-          orientation: [1, 0],
-          pos: [0, 0.5]
-        }
-      ]
+      endpoints: [{
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '4',
-      label: '条件分支',
-      className: 'orange',
-      iconType: 'icon-slbfuzaijunhengSLB',
-      top: 250,
-      left: 250,
+      label: 'Employee',
+      className: 'simple-blue',
+      iconType: 'icon-guanlian',
+      top: 125,
+      left: 520,
       Class: Node,
-      endpoints: [
-        {
-          id: 'right',
-          orientation: [1, 0],
-          pos: [0, 0.5]
-        },
-        {
-          id: 'left',
-          orientation: [-1, 0],
-          pos: [0, 0.5]
-        }
-      ]
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }, {
+        id: 'right',
+        orientation: [1, 0],
+        pos: [0, 0.5]
+      }, {
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '5',
-      label: '行内准入',
-      className: 'deep-blue',
-      iconType: 'icon-guize-kai',
-      top: 180,
-      left: 400,
+      label: 'Gets Paid',
+      className: 'green',
+      iconType: 'icon-rds',
+      top: 125,
+      left: 740,
       Class: Node,
-      endpoints: [
-        {
-          id: 'right',
-          orientation: [1, 0],
-          pos: [0, 0.5]
-        },
-        {
-          id: 'left',
-          orientation: [-1, 0],
-          pos: [0, 0.5]
-        }
-      ]
+      endpoints: [{
+        id: 'right',
+        orientation: [1, 0],
+        pos: [0, 0.5]
+      }, {
+        id: 'left',
+        orientation: [-1, 0],
+        pos: [0, 0.5]
+      }]
     },
     {
       id: '6',
-      label: '拒绝',
+      label: 'Wage',
       className: 'simple-blue',
-      iconType: 'icon-rds',
-      top: 330,
-      left: 400,
+      iconType: 'icon-guanlian',
+      top: 125,
+      left: 960,
       Class: Node,
-      endpoints: [
-        {
-          id: 'right',
-          orientation: [1, 0],
-          pos: [0, 0.5]
-        },
-        {
-          id: 'left',
-          orientation: [-1, 0],
-          pos: [0, 0.5]
-        }
-      ]
+      endpoints: [{
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }, {
+        id: 'left',
+        orientation: [-1, 0],
+        pos: [0, 0.5]
+      }]
     },
     {
       id: '7',
-      label: '结束',
-      className: 'gray',
-      iconType: 'icon-tingzhi',
-      top: 330,
-      left: 550,
+      label: 'Amount',
+      className: 'purple',
+      iconType: 'icon-wo',
+      top: 225,
+      left: 800,
       Class: Node,
-      endpoints: [
-        {
-          id: 'left',
-          orientation: [-1, 0],
-          pos: [0, 0.5]
-        }
-      ]
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '8',
-      label: '企业资质评分卡',
-      className: 'simple-yellow',
-      iconType: 'icon-rds',
-      top: 180,
-      left: 750,
+      label: 'Date',
+      className: 'purple',
+      iconType: 'icon-wo',
+      top: 225,
+      left: 1120,
       Class: Node,
-      endpoints: [
-        {
-          id: 'left',
-          orientation: [-1, 0],
-          pos: [0, 0.5]
-        }
-      ]
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '9',
-      label: '企业评分卡等级',
-      className: 'deep-green',
-      iconType: 'icon-ossduixiangcunchuOSS',
-      top: 270,
-      left: 750,
-      Class: Node
+      label: 'Salesman',
+      className: 'simple-blue',
+      iconType: 'icon-guanlian',
+      top: 275,
+      left: 520,
+      Class: Node,
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }, {
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '10',
-      label: '条件分支',
-      className: 'orange',
-      iconType: 'icon-slbfuzaijunhengSLB',
-      top: 360,
-      left: 771,
+      label: 'User',
+      className: 'green',
+      iconType: 'icon-guanlian',
+      top: 425,
+      left: 520,
       Class: Node,
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }, {
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '11',
-      label: 'E档拒绝',
+      label: 'Company',
       className: 'simple-blue',
-      iconType: 'icon-rds',
-      top: 460,
-      left: 700,
-      Class: Node
+      iconType: 'icon-guanlian',
+      top: 575,
+      left: 520,
+      Class: Node,
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }, {
+        id: 'bottom',
+        orientation: [0, 1],
+        pos: [0.5, 0]
+      }]
     },
     {
       id: '12',
-      label: '结束',
-      className: 'gray',
-      iconType: 'icon-tingzhi',
-      top: 540,
-      left: 710,
-      Class: Node
-    },
-    {
-      id: '13',
-      label: '企业信用额度',
+      label: 'Plate',
       className: 'purple',
-      iconType: 'icon-shujuji',
-      top: 460,
-      left: 850,
+      iconType: 'icon-wo',
+      top: 725,
+      left: 520,
       Class: Node,
-    },
-    {
-      id: '14',
-      label: '回落脚本',
-      className: 'pink',
-      iconType: 'icon-edit',
-      top: 540,
-      left: 864,
-      Class: Node
-    },
-    {
-      id: '15',
-      label: '赋值通过',
-      className: 'simple-blue',
-      iconType: 'icon-rds',
-      top: 620,
-      left: 864,
-      Class: Node
-    },
-    {
-      id: '16',
-      label: '结束',
-      className: 'gray',
-      iconType: 'icon-bofang',
-      top: 700,
-      left: 877,
-      Class: Node
+      endpoints: [{
+        id: 'top',
+        orientation: [0, -1],
+        pos: [0.5, 0]
+      }]
     },
   ],
   edges: [
     {
-      source: '1',
-      target: '2',
+      source: 'bottom',
+      target: 'top',
       sourceNode: '1',
-      targetNode: '2',
+      targetNode: '4',
+      type: 'endpoint',
       arrow: true,
-      arrowPosition: 0.25,
+      arrowPosition: 0.5,
       Class: Edge
     },
     {
-      source: '2',
-      target: '3',
+      source: 'bottom',
+      target: 'top',
       sourceNode: '2',
-      targetNode: '3',
+      targetNode: '4',
+      type: 'endpoint',
       arrow: true,
-      arrowPosition: 0.25,
+      arrowPosition: 0.5,
       Class: Edge
     },
     {
-      source: 'right',
-      target: 'left',
+      source: 'bottom',
+      target: 'top',
       sourceNode: '3',
       targetNode: '4',
       type: 'endpoint',
       arrow: true,
-      arrowPosition: 0.1,
+      arrowPosition: 0.5,
       Class: Edge
     },
     {
       source: 'right',
       target: 'left',
+      label: 1,
       sourceNode: '4',
       targetNode: '5',
       type: 'endpoint',
-      arrow: true,
-      arrowPosition: 0.1,
-      Class: Edge
-    },
-    {
-      source: 'right',
-      target: 'left',
-      sourceNode: '4',
-      targetNode: '6',
-      type: 'endpoint',
-      arrow: true,
-      arrowPosition: 0.1,
-      Class: Edge
-    },
-    {
-      source: 'right',
-      target: 'left',
-      sourceNode: '6',
-      targetNode: '7',
-      type: 'endpoint',
-      arrow: true,
-      arrowPosition: 0.1,
-      Class: Edge
-    },
-    {
-      source: 'right',
-      target: 'left',
-      sourceNode: '5',
-      targetNode: '8',
-      type: 'endpoint',
-      arrow: true,
-      arrowPosition: 0.03,
-      Class: Edge
-    },
-    {
-      source: '8',
-      target: '9',
-      sourceNode: '8',
-      targetNode: '9',
-      arrow: true,
-      arrowPosition: 0.25,
-      Class: Edge
-    },
-    {
-      source: '9',
-      target: '10',
-      sourceNode: '9',
-      targetNode: '10',
-      arrow: true,
-      arrowPosition: 0.25,
-      Class: Edge
-    },
-    {
-      source: '10',
-      target: '11',
-      sourceNode: '10',
-      targetNode: '11',
       arrow: true,
       arrowPosition: 0.2,
       Class: Edge
     },
     {
-      source: '11',
-      target: '12',
+      source: 'right',
+      target: 'left',
+      label: 'N',
+      sourceNode: '5',
+      targetNode: '6',
+      type: 'endpoint',
+      arrow: true,
+      arrowPosition: 0.2,
+      Class: Edge
+    },
+    {
+      source: 'bottom',
+      target: 'top',
+      sourceNode: '6',
+      targetNode: '7',
+      type: 'endpoint',
+      arrow: true,
+      arrowPosition: 0.5,
+      Class: Edge
+    },
+    {
+      source: 'bottom',
+      target: 'top',
+      sourceNode: '6',
+      targetNode: '8',
+      type: 'endpoint',
+      arrow: true,
+      arrowPosition: 0.5,
+      Class: Edge
+    },
+    {
+      source: 'bottom',
+      target: 'top',
+      label: 'ISA',
+      sourceNode: '4',
+      targetNode: '9',
+      type: 'endpoint',
+      arrow: true,
+      arrowPosition: 0.2,
+      Class: Edge
+    },
+    {
+      source: 'bottom',
+      target: 'top',
+      label: '0..1',
+      sourceNode: '9',
+      targetNode: '10',
+      type: 'endpoint',
+      arrow: true,
+      arrowPosition: 0.2,
+      Class: Edge
+    },
+    {
+      source: 'bottom',
+      target: 'top',
+      label: '1..1',
+      sourceNode: '10',
+      targetNode: '11',
+      type: 'endpoint',
+      arrow: true,
+      arrowPosition: 0.2,
+      Class: Edge
+    },
+    {
+      source: 'bottom',
+      target: 'top',
       sourceNode: '11',
       targetNode: '12',
+      type: 'endpoint',
       arrow: true,
-      arrowPosition: 0.3,
-      Class: Edge
-    },
-    {
-      source: '10',
-      target: '13',
-      sourceNode: '10',
-      targetNode: '13',
-      arrow: true,
-      arrowPosition: 0.18,
-      Class: Edge
-    },
-    {
-      source: '13',
-      target: '14',
-      sourceNode: '13',
-      targetNode: '14',
-      arrow: true,
-      arrowPosition: 0.3,
-      Class: Edge
-    },
-    {
-      source: '14',
-      target: '15',
-      sourceNode: '14',
-      targetNode: '15',
-      arrow: true,
-      arrowPosition: 0.3,
-      Class: Edge
-    },
-    {
-      source: '15',
-      target: '16',
-      sourceNode: '15',
-      targetNode: '16',
-      arrow: true,
-      arrowPosition: 0.3,
+      arrowPosition: 0.5,
       Class: Edge
     },
   ]
