+++ example/demo/decision-new/index.less
@@ -1,11 +1,16 @@
-.decision-page-new {
+.decision-new {
   width: 100%;
   height: 100%;
-  .decision-canvas {
+  .decision-new-canvas {
     width: 100%;
     height: 100%;
-    .butterflie-circle-endpoint {
-      display: none;
-    }
+  }
+  .label {
+    position: absolute;
+    width: 38px;
+    height: 22px;
+    background: #EBEBEB;
+    text-align: center;
+    border-radius: 100px;
   }
 }
