+++ example/demo/decision-new/node.less
@@ -1,7 +1,7 @@
-.decision-base-node {
+.decision-new-node {
   position: absolute;
   height: 36px;
-  min-width: 80px;
+  width: 116px;
   line-height: 36px;
   text-align: center;
   background: #fff;
@@ -22,10 +22,12 @@
   }
   .text-box {
     display: inline-block;
-    height: 30px;
-    min-width: 50px;
+    height: 36px;
+    width: 84px;
     padding: 0 10px;
-    line-height: 30px;
+    line-height: 36px;
+    font-size: 12px;
+    vertical-align: bottom;
   }
   .green {
     background: #73D13D;
