+++ example/demo/entityRelationship-new/index.less
@@ -20,7 +20,7 @@
     background-color: #FFA940;
   }
 
-  .Orange {
+  .orange {
     background-color: #FFA940;
   }
