+++ example/demo/policy-new/index.less
@@ -1,0 +1,11 @@
+.policy-page-new {
+  width: 100%;
+  height: 100%;
+  .policy-canvas {
+    width: 100%;
+    height: 100%;
+    .butterflie-circle-endpoint {
+      display: none;
+    }
+  }
+}
