+++ example/demo/policy-new/node.less
@@ -1,7 +1,7 @@
-.scene6-new-node {
+.policy-base-node {
   position: absolute;
   height: 36px;
-  width: 116px;
+  min-width: 80px;
   line-height: 36px;
   text-align: center;
   background: #fff;
@@ -22,12 +22,10 @@
   }
   .text-box {
     display: inline-block;
-    height: 36px;
-    width: 84px;
+    height: 30px;
+    min-width: 50px;
     padding: 0 10px;
-    line-height: 36px;
-    font-size: 12px;
-    vertical-align: bottom;
+    line-height: 30px;
   }
   .green {
     background: #73D13D;
