--- example/demo/scene-6-new/index.less
@@ -1,16 +1,0 @@
-.scene-6-new {
-  width: 100%;
-  height: 100%;
-  .scene-6-new-canvas {
-    width: 100%;
-    height: 100%;
-  }
-  .label {
-    position: absolute;
-    width: 38px;
-    height: 22px;
-    background: #EBEBEB;
-    text-align: center;
-    border-radius: 100px;
-  }
-}
