+++ example/demo/schedule-new/index.less
@@ -1,7 +1,7 @@
-.scene-5-new {
+.schedule-new {
   width: 100%;
   height: 100%;
-  .scene-5-new-canvas {
+  .schedule-new-canvas {
     width: 100%;
     height: 100%;
   }
