+++ example/demo/schedule-new/node.less
@@ -1,4 +1,4 @@
-.scene-5-base-node {
+.schedule-base-node {
   position: absolute;
   height: 36px;
   width: 116;
